+++ tests/test_wikipron/test_extract.py
@@ -2,7 +2,7 @@ import pytest
 import requests_html
 
 from wikipron.extract import EXTRACTION_FUNCTIONS
-from wikipron.extract.core import _skip_pron
+from wikipron.extract.core import _handle_parens, _skip_pron
 from wikipron.extract.default import extract_word_pron_default
 
 
@@ -36,3 +36,14 @@ def test_extraction_functions_have_the_same_signature(func):
 )
 def test__skip_pron(pron, iso639_key, skip_spaces, expected):
     assert _skip_pron(pron, skip_spaces) == expected
+
+
+@pytest.mark.parametrize(
+    "input_pron, skip_parens, expected_pron",
+    [
+        ("mɪskæɹəktəɹ(a)ɪzeɪʃən", True, "mɪskæɹəktəɹaɪzeɪʃən"),
+        ("mɪskæɹəktəɹ(a)ɪzeɪʃən", False, "mɪskæɹəktəɹ(a)ɪzeɪʃən"),
+    ],
+)
+def test__handle_parens(input_pron, skip_parens, expected_pron):
+    assert _handle_parens(input_pron, skip_parens) == expected_pron