+++ CHANGELOG.md
@@ -80,6 +80,7 @@ Unreleased
 -   Configured scripts for Kazakh (`kaz`). (\#345)
 -   Added Easten Lawa (`lwl`). (\#346)
 -   Configuration for Western Lawa (`lcp`). (\#347)
+-   Added Nyahkur (`cbn`). (\#348)
 
 #### Changed
 
@@ -151,6 +152,7 @@ Unreleased
 -   Added Vietnamese contour tone grouping test in `tests/test_config.py` (\#308)
 -   Added restart functionality. (\#340)
 -   Added very basic API for script detection. (\#341)
+-   Added `--skip-parens` and `--no-skip-parens` flags. (\#343)
 
 #### Changed